RatingBarSvg.kt is rewritten in kotlin style, isCompensatingMarginActive attribute added.
