+++ app/src/main/res/values/attrs.xml
@@ -4,6 +4,7 @@
         <attr name="innerPadding" format="dimension" />
         <attr name="drawableHeight" format="dimension" />
         <attr name="drawableWidth" format="dimension" />
+        <attr name="isCompensatingMarginActive" format="boolean" />
     </declare-styleable>
 
 </resources>
